@@ -1,45 +1,33 @@
 {
-  "name": "web",
+  "name": "rota-oc",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.8.0",
-    "@types/react-modal": "^3.13.1",
-    "@unform/core": "^2.1.6",
-    "@unform/web": "^2.1.6",
-    "@vitejs/plugin-react": "^2.1.0",
-    "axios": "^0.19.2",
-    "date-fns": "^2.16.1",
-    "ip": "^1.1.8",
-    "md5": "^2.3.0",
-    "nodejs-base64": "^2.0.0",
-    "polished": "^3.6.5",
-    "react": "^16.9.0",
-    "react-dom": "^16.8.0",
-    "react-hook-form": "^7.35.0",
-    "react-icons": "^3.10.0",
-    "react-modal": "^3.15.1",
-    "react-router-dom": "^5.2.0",
-    "react-select": "^5.4.0",
-    "react-spring": "^8.0.27",
-    "styled-components": "^5.1.1",
-    "typescript": "^4.8.2",
-    "uuidv4": "^6.1.1",
-    "vite": "^3.1.8",
-    "vite-plugin-svgr": "^2.2.2",
-    "vite-tsconfig-paths": "^3.5.1",
-    "yup": "^0.29.1"
+    "@headlessui/react": "^1.7.3",
+    "@heroicons/react": "^2.0.12",
+    "@tailwindcss/forms": "^0.5.3",
+    "@types/node": "^18.11.5",
+    "@types/react": "^18.0.22",
+    "@types/react-dom": "^18.0.7",
+    "autoprefixer": "^10.4.12",
+    "postcss": "^8.4.18",
+    "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6.4.2",
+    "tailwindcss": "^3.2.1",
+    "typescript": "^4.8.4"
   },
   "scripts": {
-    "start": "vite",
+    "dev": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -54,20 +42,10 @@
     ]
   },
   "devDependencies": {
-    "@types/react-router-dom": "^5.1.5",
-    "@types/styled-components": "^5.1.1",
-    "@types/yup": "^0.29.3",
-    "@typescript-eslint/eslint-plugin": "^3.6.0",
-    "@typescript-eslint/parser": "^3.6.0",
-    "eslint": "6.8.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^2.5.0",
-    "prettier": "^2.0.5"
+    "@types/react-router-dom": "5.3.3",
+    "@vitejs/plugin-react": "^2.1.0",
+    "vite": "^3.1.8",
+    "vite-plugin-svgr": "^2.2.2",
+    "vite-tsconfig-paths": "^3.5.2"
   }
 }
